+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
+- AnyButtonPressedListener.waitingTime
 - DisableCursorListener.OnVisibilityChanged event
 
 ## [1.3.0] - 2025-09-06